# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * A3 Sales OS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface Lead {
  id: number;
  pipelineType: string;
  companyName: string;
  contactName: string;
  title?: string;
  email?: string;
  phone?: string;
  linkedin?: string;
  location?: string;
  industry?: string;
  venueProperty?: string;
  projectType?: string;
  estimatedBudget?: number;
  status: string;
  lastContactDate?: string;
  nextStep?: string;
  nextFollowUpDate?: string;
  notes?: string;
  dealValueEstimate?: number;
  proposalValue?: number;
  closeProbability?: number;
  forecastValue?: number;
  source?: string;
  createdAt: string;
  updatedAt: string;
}

export interface LeadInput {
  pipelineType: string;
  companyName: string;
  contactName: string;
  title?: string;
  email?: string;
  phone?: string;
  linkedin?: string;
  location?: string;
  industry?: string;
  venueProperty?: string;
  projectType?: string;
  estimatedBudget?: number;
  status: string;
  lastContactDate?: string;
  nextStep?: string;
  nextFollowUpDate?: string;
  notes?: string;
  dealValueEstimate?: number;
  proposalValue?: number;
  closeProbability?: number;
  source?: string;
}

export interface Task {
  id: number;
  leadId?: number;
  taskType: string;
  dueDate?: string;
  status: string;
  notes?: string;
  createdAt: string;
  updatedAt: string;
  leadCompanyName?: string;
  leadContactName?: string;
}

export interface TaskInput {
  leadId?: number;
  taskType: string;
  dueDate?: string;
  status: string;
  notes?: string;
}

export interface Template {
  id: number;
  name: string;
  category: string;
  subject?: string;
  body: string;
  createdAt: string;
  updatedAt: string;
}

export interface TemplateInput {
  name: string;
  category: string;
  subject?: string;
  body: string;
}

export interface Asset {
  id: number;
  title: string;
  category: string;
  description?: string;
  url?: string;
  linkedLeadId?: number;
  createdAt: string;
  updatedAt: string;
}

export interface AssetInput {
  title: string;
  category: string;
  description?: string;
  url?: string;
  linkedLeadId?: number;
}

export type DashboardDataPipelineByStageItem = {
  stage: string;
  count: number;
  value: number;
};

export type DashboardDataPipelineByTypeItem = {
  type: string;
  count: number;
  value: number;
};

export interface DashboardData {
  totalLeads: number;
  newLeads: number;
  contacted: number;
  replied: number;
  qualified: number;
  meetingsBooked: number;
  meetingsCompleted: number;
  proposalsSent: number;
  negotiation: number;
  closedWon: number;
  closedLost: number;
  nurture: number;
  totalPipelineValue: number;
  totalForecastValue: number;
  closedRevenue: number;
  followUpsDueToday: number;
  overdueFollowUps: number;
  pipelineByStage: DashboardDataPipelineByStageItem[];
  pipelineByType: DashboardDataPipelineByTypeItem[];
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  leadId?: number;
  createdAt: string;
}

export type GetLeadsParams = {
  search?: string;
  pipelineType?: string;
  status?: string;
  projectType?: string;
  source?: string;
  location?: string;
  followUpDueToday?: boolean;
  overdue?: boolean;
};

export type UpdateLeadStatusBody = {
  status: string;
};

export type ImportLeadsBody = {
  leads: LeadInput[];
};

export type ImportLeads200 = {
  imported: number;
  skipped: number;
  total: number;
};

export type GetTasksParams = {
  leadId?: number;
  status?: string;
  dueFilter?: GetTasksDueFilter;
};

export type GetTasksDueFilter =
  (typeof GetTasksDueFilter)[keyof typeof GetTasksDueFilter];

export const GetTasksDueFilter = {
  today: "today",
  overdue: "overdue",
  next7days: "next7days",
} as const;

export type GetTemplatesParams = {
  category?: string;
};

export type GetAssetsParams = {
  category?: string;
  search?: string;
};