# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: A3 Sales OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      summary: Get all leads
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: pipelineType
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: projectType
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: followUpDueToday
          in: query
          schema:
            type: boolean
        - name: overdue
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a single lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    put:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /leads/{id}/duplicate:
    post:
      operationId: duplicateLead
      tags: [leads]
      summary: Duplicate a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}/status:
    patch:
      operationId: updateLeadStatus
      tags: [leads]
      summary: Quick status update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
              required:
                - status
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/import:
    post:
      operationId: importLeads
      tags: [leads]
      summary: Import leads from CSV data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                leads:
                  type: array
                  items:
                    $ref: "#/components/schemas/LeadInput"
              required:
                - leads
      responses:
        "200":
          description: Import results
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
                  skipped:
                    type: integer
                  total:
                    type: integer
                required:
                  - imported
                  - skipped
                  - total
...
  /assets:
    get:
      operationId: getAssets
      tags: [assets]
      summary: Get all assets
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Create a new asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Created asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    put:
      operationId: updateAsset
      tags: [assets]
      summary: Update an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "200":
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Lead:
      type: object
      properties:
        id:
          type: integer
        pipelineType:
          type: string
        companyName:
          type: string
        contactName:
          type: string
        title:
          type: string
        email:
          type: string
        phone:
          type: string
        linkedin:
          type: string
        location:
          type: string
        industry:
          type: string
        venueProperty:
          type: string
        projectType:
          type: string
        estimatedBudget:
          type: number
        status:
          type: string
        lastContactDate:
          type: string
        nextStep:
          type: string
        nextFollowUpDate:
          type: string
        notes:
          type: string
        dealValueEstimate:
          type: number
        proposalValue:
          type: number
        closeProbability:
          type: number
        forecastValue:
          type: number
        source:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - pipelineType
        - companyName
        - contactName
        - status
        - createdAt
        - updatedAt
    LeadInput:
      type: object
      properties:
        pipelineType:
          type: string
        companyName:
          type: string
        contactName:
          type: string
        title:
          type: string
        email:
          type: string
        phone:
          type: string
        linkedin:
          type: string
        location:
          type: string
        industry:
          type: string
        venueProperty:
          type: string
        projectType:
          type: string
        estimatedBudget:
          type: number
        status:
          type: string
        lastContactDate:
          type: string
        nextStep:
          type: string
        nextFollowUpDate:
          type: string
        notes:
          type: string
        dealValueEstimate:
          type: number
        proposalValue:
          type: number
        closeProbability:
          type: number
        source:
          type: string
      required:
        - pipelineType
        - companyName
        - contactName
        - status
...
    Asset:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        description:
          type: string
        url:
          type: string
        linkedLeadId:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - category
        - createdAt
        - updatedAt
    AssetInput:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string
        url:
          type: string
        linkedLeadId:
          type: integer
      required:
        - title
        - category
...