# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * A3 Sales OS API
 * OpenAPI spec version: 0.1.0
 */

export interface Asset {
  id: number;
  title: string;
  category: string;
  description?: string;
  url?: string;
  linkedLeadId?: number;
  createdAt: string;
  updatedAt: string;
}