# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * A3 Sales OS API
 * OpenAPI spec version: 0.1.0
 */

export interface LeadInput {
  pipelineType: string;
  companyName: string;
  contactName: string;
  title?: string;
  email?: string;
  phone?: string;
  linkedin?: string;
  location?: string;
  industry?: string;
  venueProperty?: string;
  projectType?: string;
  estimatedBudget?: number;
  status: string;
  lastContactDate?: string;
  nextStep?: string;
  nextFollowUpDate?: string;
  notes?: string;
  dealValueEstimate?: number;
  proposalValue?: number;
  closeProbability?: number;
  source?: string;
}